# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/groups/{id}": {
      "get": {
        "operationId": "GroupController_get",
        "summary": "Get group details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedGroupDetailsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "GroupController_update",
        "summary": "Update a group",
        "description": "Name, description, and terms can be updated anytime. Core fields (contributionAmount, frequency, maxSlots, serviceCharge, lateFee) can only be updated before the group starts.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Group updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedGroupDetailsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "GroupController_delete",
        "summary": "Delete a NOT_STARTED group (admin only)",
        "description": "Deletes a NOT_STARTED group and all contributors. Hard delete — no audit trail needed.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "reason": {
                    "type": "string",
                    "example": "No longer needed"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Group deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedOkResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/groups/{groupId}/join-requests": {
      "post": {
        "operationId": "GroupJoinController_request",
        "summary": "Request to join a group",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "description": "The group ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Join request created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedIdResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/groups/{groupId}/join-requests/{requestId}": {
      "patch": {
        "operationId": "GroupJoinController_updateStatus",
        "summary": "Approve or reject a join request",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "description": "The group ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "requestId",
            "required": true,
            "in": "path",
            "description": "The join request ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "APPROVED",
                      "REJECTED"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Join request updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedOkResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/join/{token}": {
      "post": {
        "operationId": "JoinLinkController_consume",
        "summary": "Join a group using a join link token",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "description": "The join link token",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully joined the group",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WrappedOkResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/tickets/groups/{groupId}": {
      "get": {
        "operationId": "TicketController_getGroupTickets",
        "summary": "Get all tickets for a group (admin only)",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search query to filter results",
            "schema": {
              "example": "john",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order: asc or desc",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor for cursor-based paging",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PENDING",
                "UNDER_REVIEW",
                "APPROVED",
                "REJECTED",
                "CLOSED"
              ],
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "CONTRIBUTOR_REPLACEMENT",
                "LEAVE_GROUP",
                "DISPUTE",
                "OTHER"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group tickets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedGroupTicketsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Tickets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Hajor API",
    "description": "API documentation for Hajor fintech system",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "OkResponseDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
      "WrappedOkResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "timestamp": {
            "type": "string",
            "example": "2026-04-06T00:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/endpoint"
          },
          "requestId": {
            "type": "string",
            "nullable": true,
            "example": "uuid-here"
          },
          "data": {
            "$ref": "#/components/schemas/OkResponseDto"
          },
          "code": {
            "type": "string",
            "example": "OK"
          }
        },
        "required": [
          "statusCode",
          "timestamp",
          "path",
          "data",
          "code"
        ]
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "pages": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "pages"
        ]
      },
...
      "GroupDetailsResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "adminId": {
            "type": "string"
          },
          "terms": {
            "type": "string",
            "description": "Group terms text, if any"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "adminId"
        ]
      },
      "WrappedGroupDetailsResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "timestamp": {
            "type": "string",
            "example": "2026-04-06T00:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/endpoint"
          },
          "requestId": {
            "type": "string",
            "nullable": true,
            "example": "uuid-here"
          },
          "data": {
            "$ref": "#/components/schemas/GroupDetailsResponseDto"
          },
          "code": {
            "type": "string",
            "example": "OK"
          }
        },
        "required": [
          "statusCode",
          "timestamp",
          "path",
          "data",
          "code"
        ]
      },
...
      "UpdateGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "example": "My Savings Group"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "example": "Updated description"
          },
          "terms": {
            "type": "string",
            "maxLength": 5000,
            "example": "Updated terms..."
          },
          "contributionAmount": {
            "type": "number",
            "minimum": 0,
            "example": 5000
          },
          "frequency": {
            "type": "string",
            "enum": [
              "WEEKLY",
              "MONTHLY"
            ],
            "example": "WEEKLY"
          },
          "maxSlots": {
            "type": "number",
            "minimum": 2,
            "example": 10
          },
          "serviceCharge": {
            "type": "number",
            "minimum": 0,
            "example": 0
          },
          "lateFee": {
            "type": "number",
            "minimum": 0,
            "example": 0
          }
        }
      },
...
      "IdResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
      "WrappedIdResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "timestamp": {
            "type": "string",
            "example": "2026-04-06T00:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/endpoint"
          },
          "requestId": {
            "type": "string",
            "nullable": true,
            "example": "uuid-here"
          },
          "data": {
            "$ref": "#/components/schemas/IdResponseDto"
          },
          "code": {
            "type": "string",
            "example": "OK"
          }
        },
        "required": [
          "statusCode",
          "timestamp",
          "path",
          "data",
          "code"
        ]
      },
...
      "TicketUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email"
        ]
      },
...
      "GroupTicketResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "groupId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "contributorId": {
            "type": "string"
          },
          "newUserId": {
            "type": "string"
          },
          "adminNotes": {
            "type": "string"
          },
          "resolvedAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/TicketUser"
          }
        },
        "required": [
          "id",
          "type",
          "groupId",
          "userId",
          "status",
          "createdAt",
          "updatedAt",
          "user"
        ]
      },
      "PaginatedGroupTicketsResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupTicketResponseDto"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "required": [
          "items",
          "pagination"
        ]
      },
...
    }
  }
}